requirements.txt
django
python-telegram-bot
requests